Added separate tabs for every plot. isolation of code for every tab. Added decorators to help with file validation and exception handling. Added utils for helpers.py and caching.py. Updated visualization.py for figures. Added config.py for native default configurations.
